+++ test/cocoTestApp.js
@@ -14,6 +14,10 @@ var CocoApplication = dejavu.Class.declare({
   initialize() {
     console.log("CocoTestApp initialized, look at available Coco-Classes: ", Coco);
 
+    Coco.Translator.addEventListener(Coco.TranslatorEvent.CHANGE_LOCALE, (event) => {
+      console.info("locale changed " + event.locale);
+    });
+
     Coco.Translator.loadMessagesFromObject({
       title:  {
         1: "bla-de",